+++ sample/test-rules/invitesC.test.js
@@ -4,10 +4,10 @@
 import { strict as assert } from 'assert'
 
 //import { dbAuth } from 'firebase-jest-testing/firestoreTestingReadOnly';
-import { dbAuth } from '../src/firestoreTesting/readOnly.js';
+import { dbAuth } from '../../src/firestoreTesting/readOnly.js';
 
 //import { FieldValue } from 'firebase-jest-testing/firestoreTesting';
-import { FieldValue } from '../src/firestoreTesting/index.js';
+import { FieldValue } from '../../src/firestoreTesting/index.js';
 
 import { test, expect, describe, beforeAll } from '@jest/globals'