+++ sample/test-rules/projectsC.test.js
@@ -5,10 +5,10 @@ import { strict as assert } from 'assert'
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
 //import { dbAuth } from 'firebase-jest-testing/firestoreTestingReadOnly';
-import { dbAuth } from '../src/firestoreTesting/readOnly.js'
+import { dbAuth } from '../../src/firestoreTesting/readOnly.js'
 
 //import { FieldValue } from 'firebase-jest-testing/firestoreTesting';
-import { FieldValue } from '../src/firestoreTesting/index.js'
+import { FieldValue } from '../../src/firestoreTesting/index.js'
 
 const anyDate = new Date();   // a non-server date