+++ sample/test-rules/userInfoC.test.js
@@ -5,7 +5,7 @@ import { strict as assert } from 'assert'
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
 //import { dbAuth } from 'firebase-jest-testing/firestoreTestingReadOnly';
-import { dbAuth } from '../src/firestoreTesting/readOnly.js';
+import { dbAuth } from '../../src/firestoreTesting/readOnly.js';
 
 describe("'/userInfo' rules", () => {
   let unauth_userinfoC, abc_userinfoC, def_userinfoC;