+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 16-Aug-20
+
+- reorganisation to `sample/test-fns`, `sample/test-rules`
+- [x] `npm run test:rest:all` passes
+- [ ] `npm run test:rules:all` passes
+
 ## 14-Aug-20
 
 - Security Rules tests implemented (uses ES modules, yay!)